requirements.txt
Flask
pymongo
boto3
pydicom
Pillow
python-dotenv
